+++ lib/objectmancy.rb
@@ -1,4 +1,5 @@
 require 'objectmancy/objectable'
+require 'objectmancy/hashable'
 
 # Namespace for all Objectmancy functionality
 # Also serves as a mixin to include Objectable